Added missing imports to get module system working
I hadn't been exporting the Matrix class from MLToolkit,
so the `importall MLToolkit` in the inner modules hadn't
been importing the definition of Matrix, so all my train
functions in the inner modules were defining train with
Base.Matrix instead of MLToolkit.Matrix. This fixes that,
so each learner can be in it's own module to not worry
about naming constants and functions uniquely.