Updated a ton of stuff
Changed the Matrix class quite a bit to take advantage
of some features in Julia that I hadn't been taking advantage of.
The changes were also intended to make the Matrix function calls
and the implementation of the Matrix functions more intuitive.

Also added support for passing keyword arguments to learners from
the command line. To pass the argument from the command line,
the name should be preceded by two dashes, followed by a space,
and followed by the argument value. See the BaselineLearner for
an example of how to add keyword arguments to the constructor
of a learner.